BEO-5797 Playback_adapter use  LayeredCache

+++ ApiObjects/Rules/RuleAction.cs
@@ -119,7 +119,7 @@ namespace ApiObjects.Rules
     public class ApplyPlaybackAdapterRuleAction : AssetRuleAction
     {
         [JsonProperty("AdapterId")]
-        public long AdapterId { get; set; }
+        public int AdapterId { get; set; }
 
         public ApplyPlaybackAdapterRuleAction()
         {
